# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VA eBay Client Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ebay-accounts:
    get:
      operationId: getEbayAccounts
      tags: [ebay-accounts]
      summary: List all eBay accounts
      parameters:
        - name: clientId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of eBay accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EbayAccount"
    post:
      operationId: createEbayAccount
      tags: [ebay-accounts]
      summary: Create eBay account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEbayAccountBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EbayAccount"
  /ebay-accounts/{id}:
    get:
      operationId: getEbayAccount
      tags: [ebay-accounts]
      summary: Get eBay account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: eBay account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EbayAccount"
        "404":
          description: Not found
    patch:
      operationId: updateEbayAccount
      tags: [ebay-accounts]
      summary: Update eBay account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEbayAccountBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EbayAccount"
    delete:
      operationId: deleteEbayAccount
      tags: [ebay-accounts]
      summary: Delete eBay account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Wise & Cards ───────────────────────────────────────────────────────────
...
  /invoices:
    get:
      operationId: getInvoices
      tags: [invoices]
      summary: List all invoices
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /expenses:
    get:
      operationId: getExpenses
      tags: [expenses]
      summary: List all expenses
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExpenseBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Recovery & 2FA ─────────────────────────────────────────────────────────
  /recovery:
    get:
      operationId: getRecoveryEntries
      tags: [recovery]
      summary: List all recovery entries
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecoveryEntry"
    post:
      operationId: createRecoveryEntry
      tags: [recovery]
      summary: Create recovery entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecoveryEntryBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryEntry"
  /recovery/{id}:
    patch:
      operationId: updateRecoveryEntry
      tags: [recovery]
      summary: Update recovery entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRecoveryEntryBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryEntry"
    delete:
      operationId: deleteRecoveryEntry
      tags: [recovery]
      summary: Delete recovery entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Daily Login ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    EbayAccount:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        ebayUsername:
          type: string
        ebayEmail:
          type: ["string", "null"]
        ebayPassword:
          type: ["string", "null"]
        marketplace:
          type: ["string", "null"]
        accountType:
          type: ["string", "null"]
        accountHealth:
          type: ["string", "null"]
        feedbackScore:
          type: ["integer", "null"]
        topRated:
          type: boolean
        activeListings:
          type: ["integer", "null"]
        sellingLimit:
          type: ["number", "null"]
        mcRestriction:
          type: boolean
        lastLogin:
          type: ["string", "null"]
        lastPwdChange:
          type: ["string", "null"]
        pwdExpiresIn:
          type: ["integer", "null"]
        twoFaEnabled:
          type: boolean
        linkedBankId:
          type: ["string", "null"]
        linkedCardId:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, clientId, clientName, ebayUsername, topRated, mcRestriction, twoFaEnabled, createdAt, updatedAt]
    CreateEbayAccountBody:
      type: object
      properties:
        clientId:
          type: integer
        ebayUsername:
          type: string
        ebayEmail:
          type: ["string", "null"]
        ebayPassword:
          type: ["string", "null"]
        marketplace:
          type: ["string", "null"]
        accountType:
          type: ["string", "null"]
        accountHealth:
          type: ["string", "null"]
        feedbackScore:
          type: ["integer", "null"]
        topRated:
          type: boolean
        activeListings:
          type: ["integer", "null"]
        sellingLimit:
          type: ["number", "null"]
        mcRestriction:
          type: boolean
        lastLogin:
          type: ["string", "null"]
        lastPwdChange:
          type: ["string", "null"]
        pwdExpiresIn:
          type: ["integer", "null"]
        twoFaEnabled:
          type: boolean
        linkedBankId:
          type: ["string", "null"]
        linkedCardId:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required: [clientId, ebayUsername, topRated, mcRestriction, twoFaEnabled]
    UpdateEbayAccountBody:
      type: object
      properties:
        clientId:
          type: ["integer", "null"]
        ebayUsername:
          type: ["string", "null"]
        ebayEmail:
          type: ["string", "null"]
        ebayPassword:
          type: ["string", "null"]
        marketplace:
          type: ["string", "null"]
        accountType:
          type: ["string", "null"]
        accountHealth:
          type: ["string", "null"]
        feedbackScore:
          type: ["integer", "null"]
        topRated:
          type: ["boolean", "null"]
        activeListings:
          type: ["integer", "null"]
        sellingLimit:
          type: ["number", "null"]
        mcRestriction:
          type: ["boolean", "null"]
        lastLogin:
          type: ["string", "null"]
        lastPwdChange:
          type: ["string", "null"]
        pwdExpiresIn:
          type: ["integer", "null"]
        twoFaEnabled:
          type: ["boolean", "null"]
        linkedBankId:
          type: ["string", "null"]
        linkedCardId:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...
    Invoice:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        clientId:
          type: integer
        clientName:
          type: string
        issueDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        hours:
          type: ["number", "null"]
        rate:
          type: ["number", "null"]
        amount:
          type: number
        status:
          type: string
        paidOn:
          type: ["string", "null"]
        daysOverdue:
          type: ["integer", "null"]
        method:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, invoiceNumber, clientId, clientName, amount, status, createdAt, updatedAt]
    CreateInvoiceBody:
      type: object
      properties:
        invoiceNumber:
          type: string
        clientId:
          type: integer
        issueDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        hours:
          type: ["number", "null"]
        rate:
          type: ["number", "null"]
        amount:
          type: number
        status:
          type: string
        paidOn:
          type: ["string", "null"]
        method:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required: [invoiceNumber, clientId, amount, status]
    UpdateInvoiceBody:
      type: object
      properties:
        invoiceNumber:
          type: ["string", "null"]
        clientId:
          type: ["integer", "null"]
        issueDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        hours:
          type: ["number", "null"]
        rate:
          type: ["number", "null"]
        amount:
          type: ["number", "null"]
        status:
          type: ["string", "null"]
        paidOn:
          type: ["string", "null"]
        method:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...
    Expense:
      type: object
      properties:
        id:
          type: integer
        date:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        vendor:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        amount:
          type: number
        currency:
          type: ["string", "null"]
        recurring:
          type: boolean
        linkedClientId:
          type: ["integer", "null"]
        linkedClientName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, amount, recurring, createdAt, updatedAt]
    CreateExpenseBody:
      type: object
      properties:
        date:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        vendor:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        amount:
          type: number
        currency:
          type: ["string", "null"]
        recurring:
          type: boolean
        linkedClientId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
      required: [amount, recurring]
    UpdateExpenseBody:
      type: object
      properties:
        date:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        vendor:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        amount:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        recurring:
          type: ["boolean", "null"]
        linkedClientId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
    RecoveryEntry:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        accountService:
          type: string
        recoveryEmail:
          type: ["string", "null"]
        recoveryPhone:
          type: ["string", "null"]
        securityQ1:
          type: ["string", "null"]
        answer1:
          type: ["string", "null"]
        securityQ2:
          type: ["string", "null"]
        answer2:
          type: ["string", "null"]
        twoFaMethod:
          type: ["string", "null"]
        backupCodes:
          type: ["string", "null"]
        totpSeed:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, clientId, clientName, accountService, createdAt, updatedAt]
    CreateRecoveryEntryBody:
      type: object
      properties:
        clientId:
          type: integer
        accountService:
          type: string
        recoveryEmail:
          type: ["string", "null"]
        recoveryPhone:
          type: ["string", "null"]
        securityQ1:
          type: ["string", "null"]
        answer1:
          type: ["string", "null"]
        securityQ2:
          type: ["string", "null"]
        answer2:
          type: ["string", "null"]
        twoFaMethod:
          type: ["string", "null"]
        backupCodes:
          type: ["string", "null"]
        totpSeed:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required: [clientId, accountService]
    UpdateRecoveryEntryBody:
      type: object
      properties:
        clientId:
          type: ["integer", "null"]
        accountService:
          type: ["string", "null"]
        recoveryEmail:
          type: ["string", "null"]
        recoveryPhone:
          type: ["string", "null"]
        securityQ1:
          type: ["string", "null"]
        answer1:
          type: ["string", "null"]
        securityQ2:
          type: ["string", "null"]
        answer2:
          type: ["string", "null"]
        twoFaMethod:
          type: ["string", "null"]
        backupCodes:
          type: ["string", "null"]
        totpSeed:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...